Version 1.3
Bugfix to enable user specified mappings for m, m<Space> and m<BS>
Displays a menu of options under Plugins > Signature. Menu display can be disabled or customized as required.
Modified marker navigation mappings to be consistent with others in the use of [ and  ] to go to the prev and next respectively.
Updated MapKey function